W-12560400: bumped to ScalaJS 1.6. Added cast proposed by the ScalaJSCompiler

+++ js/src/main/scala/org/yaml/builder/JsOutputBuilder.scala
@@ -54,7 +54,7 @@ class JsOutputBuilder extends DocBuilder[js.Any] {
     case Str   => scalar.value.toString
     case Bool  => scalar.value.asInstanceOf[Boolean]
     case Float => scalar.value.asInstanceOf[Double]
-    case Int => scalar.value.asInstanceOf[Long].toInt
+    case Int => scalar.value.asInstanceOf[Long].toDouble
     case Null   => null
   }
